fix #29 - create error boundary and error component
